fix: orphan service cfg
